+++ enki/blfs/copy-to-share.sh
@@ -2,3 +2,4 @@
 
 cp -r config/* /tmp/vm-share
 cp -r scripts/* /tmp/vm-share
+cp -r sources/* /tmp/vm-share