+++ enki/blfs/scripts/010-nspr.sh
@@ -1,0 +1,59 @@
+#!/bin/bash
+
+
+if [ "$SHAREDDIR" = "" ]; then
+    echo 'Variable $SHAREDDIR should be defined'
+    exit 255
+fi
+
+PACKAGE=nspr-4.38.2
+FLAG=/lfsflags/blfs/$PACKAGE
+SOURCE=nspr-4.38.2.tar.gz
+
+if [ -f "$FLAG" ]; then
+    echo "Package $PACKAGE is already built."
+    exit 255
+fi
+
+if [ ! -f "/sources/$SOURCE" ]; then
+    cp "$SHAREDDIR/$SOURCE" /sources
+fi
+
+cd /sources
+
+if [ ! -d $PACKAGE ]; then
+  tar xvf $SOURCE
+  if [ $? -eq 1 ]; then
+      echo "unable to extract archive"
+      exit 255
+  fi
+fi
+
+cd $PACKAGE
+
+cd nspr &&
+
+sed -i '/^RELEASE/s|^|#|' pr/src/misc/Makefile.in &&
+sed -i 's|$(LIBRARY) ||'  config/rules.mk         &&
+
+./configure --prefix=/usr   \
+            --with-mozilla  \
+            --with-pthreads \
+            $([ $(uname -m) = x86_64 ] && echo --enable-64bit) &&
+make
+
+if [ "$?" != 0 ]; then
+    echo "Build of $PACKAGE failed"
+    exit 255
+fi
+
+sudo make install
+
+if [ "$?" != 0 ]; then
+    echo "Install of $PACKAGE failed"
+    exit 255
+fi
+
+touch $FLAG
+cd /sources
+rm -rf $PACKAGE