+++ enki/blfs/history.md
@@ -9,9 +9,9 @@ We will assume the LFS installation, is run under QUEMU.
 We want to be able to copy files from the host to the LFS installations. 
 To do so, we will use [virtiofsd](https://wiki.archlinux.org/title/QEMU#Host_file_sharing_with_virtiofsd).
 
-We fist need to install virtiofsd:
+We first need to install virtiofsd:
 
-    sudo pacpam -S virtiofsd
+    sudo pacman -S virtiofsd
 
 We run the virtiofs daemon:
 
@@ -33,6 +33,8 @@ We can then run the guest (LFS):
         -object memory-backend-memfd,id=mem,size=4G,share=on \
         -numa node,memdev=mem \
         -chardev socket,id=char0,path=/tmp/vm-share.sock \
+        -cpu host \
+        -enable-kvm \mkdir .lfsflags  
         -device vhost-user-fs-pci,chardev=char0,tag=myfs
 
 on the host, we give the access right to our normal user:
@@ -51,6 +53,60 @@ The scripts are in the directory `./scripts`. To use them one need to:
 * on the guest, run them as `/mnt/vm-share/...` 
 * on the guest export the variable `SHAREDDIR` with the value `/mnt/vm-share`
 
+The build scripts use files in `/lfsflags/blfs` to avoid repeating 
+the build, so we need to create that directory.
+
+## Downloading packages
+
+As the guest (LFS) isn't able to dowload packages yet, we download them 
+on the host with the script [download.sh]()./download.sh).
+
 ## Add an user
 
-One can add an user by running the script [01-create-user.sh](./scripts/01-create-user.sh)
+One can add an user by running the script [001-create-user.sh](./scripts/001-create-user.sh)
+
+## Build *sudo*
+
+To be able to effectively use the new user, one need to build `sudo`.
+
+The build script is [002-sudo.sh](./scripts/002-sudo.sh).
+
+One needs to add the user to the group `wheel`:
+
+    usermod -a -G wheel username
+
+where *username* si replacved by the actual user name.
+
+After checking that sudo actually works, one can disable
+the root password:
+
+     sudo passwd -d root
+     sudo passwd --lock root
+
+## Build some other packages
+
+As it's fun to have *fastfetch* (in extra), we fist need to build
+some packages befotre being able to build *fastfetch*:
+
+* [scripts/003-cmake.sh](003-cmake.sh)
+
+We need an editor (for those that don't like `vim`):
+
+* [scripts/004-nano.sh](004-nano.sh)
+
+We need program to download files from the internet:
+
+* [scripts/005-libunistring.sh](005-libunistring.sh)
+* [scripts/006-libidn2.sh](006-libidn2.sh)
+* [scripts/007-libpsl.sh](007-libpsl.sh)
+* [scripts/008-libtasn1.sh](008-libtasn1.sh)
+* [scripts/009-p11-kit.sh](009-p11-kit.sh)
+* [scripts/010-nspr.sh](010-nspr.sh)
+* [scripts/011-nss.sh](nss.sh)
+* [scripts/012-make-ca.sh](make-ca.sh)
+* [scripts/013-nettle.sh](nettle.sh)
+* [scripts/014-gnutls.sh](gnutls.sh)
+* [scripts/015-wget.sh](wget.sh)
+* [scripts/016-curl.sh](curl.sh)
+
+note: many wget tests failed, for unknown reason. The program seems to work however.